requirements2.txt
# Web framework e utilitários
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==2.7.2

# OCR e imagem
easyocr==1.7.1
pillow==10.3.0
PyMuPDF==1.24.5

# LangChain e integrações
langchain==0.2.5
langchain-core==0.2.5
langchain-openai==0.1.8
langchain-community==0.2.5

# SQL e banco de dados
SQLAlchemy==2.0.30
pymysql==1.1.1

# Ferramentas
black==24.4.2
